Put scripts in separate files to run normally with whelktool
