+++ whelktool/src/main/groovy/datatool/scripts/mergeworks/Util.groovy
@@ -4,9 +4,13 @@ import org.apache.commons.lang3.StringUtils
 import whelk.Whelk
 import whelk.util.Unicode
 
-import java.util.regex.Pattern
-
 class Util {
+    static def whelk = Whelk.createLoadedCoreWhelk()
+
+    static def clusters = 'clusters'
+
+    static def contributionPath = ['@graph', 1, 'instanceOf', 'contribution']
+
     static def titleComponents = ['mainTitle', 'titleRemainder', 'subtitle', 'hasPart', 'partNumber', 'partName', 'marc:parallelTitle', 'marc:equalTitle']
 
     static def titleVariant = ['Title', 'ParallelTitle']
@@ -150,7 +154,7 @@ class Util {
                 .collect { it['flatTitle'] }
     }
 
-    static String chipString(def thing, Whelk whelk) {
+    static String chipString(def thing) {
         if (thing instanceof Integer) {
             return thing
         }
@@ -221,94 +225,19 @@ class Util {
         return null
     }
 
-    static Map<String, List<Tuple2<Relator, Boolean>>> parseRespStatement(String respStatement) {
-        def parsedContributions = [:]
-
-        respStatement.split(';').eachWithIndex { part, i ->
-            // TODO: generalize for other material types
-            parseSwedishFictionContribution(StringUtils.normalizeSpace(part), i == 0).each { name, roles ->
-                parsedContributions
-                        .computeIfAbsent(name, r -> [])
-                        .addAll(roles)
-            }
+    static boolean nameMatch(Object local, Map agent) {
+        def variants = [agent] + asList(agent.hasVariant)
+        def name = {
+            Map p ->
+                (p.givenName && p.familyName)
+                        ? normalize("${p.givenName} ${p.familyName}")
+                        : p.name ? normalize("${p.name}") : null
         }
 
-        return parsedContributions
-    }
+        def localName = local instanceof Map ? name(local) : normalize(local)
 
-    private static Map<String, List<Tuple2<Relator, Boolean>>> parseSwedishFictionContribution(String contribution, boolean isFirstPart) {
-        def roleToPattern =
-                [
-                        (Relator.TRANSLATOR)         : ~/(bemynd(\w+|\.)? )?öf?v(\.|ers(\.|\p{L}+)?)( (till|från) \p{L}+)?|(till svenskan?|från \p{L}+)|svensk text/,
-                        (Relator.AUTHOR)             : ~/^(text(e[nr])?|skriven|written)/,
-                        (Relator.ILLUSTRATOR)        : ~/\bbild(er)?|ill(\.|ustr(\.|\w+)?)|\bvi(gn|nj)ett(er|ill)?|ritad/,
-                        (Relator.AUTHOR_OF_INTRO)    : ~/förord|inl(edn(\.|ing)|edd)/,
-                        (Relator.COVER_DESIGNER)     : ~/omslag/,
-                        (Relator.AUTHOR_OF_AFTERWORD): ~/efter(ord|skrift)/,
-                        (Relator.PHOTOGRAPHER)       : ~/\bfoto\w*\.?/,
-//                        (Relator.EDITOR)             : ~/red(\.(?! av)|aktör(er)?)|\bbearb(\.|\w+)?|återberättad|sammanställ\w*/,
-                ]
-
-        def rolePattern = ~/((?iu)${roleToPattern.values().join('|')})/
-        def followsRolePattern = ~/(:| a[fv]| by) /
-        def initialPattern = ~/\p{Lu}/
-        def namePattern = ~/\p{Lu}:?\p{Ll}+('\p{Ll})?(,? [Jj](r|unior))?/
-        def betweenNamesPattern = ~/-| |\. ?| (de(l| la)?|von|van( de[nr])?|v\.|le|af|du|dos) | [ODdLl]'/
-        def fullNamePattern = ~/(($initialPattern|$namePattern)($betweenNamesPattern)?)*$namePattern/
-        def conjPattern = ~/ (och|&|and) /
-        def roleAfterNamePattern = ~/( ?\(($rolePattern$conjPattern)?$rolePattern\))/
-        def fullContributionPattern = ~/(($rolePattern($conjPattern|\/))*$rolePattern$followsRolePattern)?$fullNamePattern($conjPattern$fullNamePattern)*$roleAfterNamePattern?/
-
-        // Make roles lower case so that they can't be mistaken for names
-        contribution = (contribution =~ rolePattern)*.first()
-                .collectEntries { [it, it.toLowerCase()] }
-                .with { contribution.replace(it) }
-
-        def nameToRoles = [:]
-
-        def matched = (contribution =~ fullContributionPattern)*.first()
-
-        matched.each { m ->
-            // Extract roles from the contribution
-            def roles = roleToPattern
-                    .findAll { k, v -> m =~ /(?iu)$v/ }
-                    .with {
-                        it.isEmpty() && contribution =~ /.+$followsRolePattern/
-                                ? [new Tuple2(Relator.UNSPECIFIED_CONTRIBUTOR, isFirstPart)]
-                                : it.collect { role, pattern -> new Tuple2(role, isFirstPart) }
-                    }
-
-            // Author should be the role if first part of respStatement (before ';') and no role seems to be stated
-            if (roles.isEmpty() && isFirstPart) {
-                roles << new Tuple2(Relator.AUTHOR, isFirstPart)
-            }
-
-            // Extract names from the contribution
-            def names = parseNames(fullNamePattern, conjPattern, m)
-
-            // Assign the roles to each name
-            nameToRoles.putAll(names.collectEntries { [it, roles] })
+        localName && variants.any {
+            name(it) && localName == name(it)
         }
-
-        return nameToRoles
-    }
-
-    private static List<String> parseNames(Pattern namePattern, Pattern conjPattern, String s) {
-        def names = []
-
-        (s =~ namePattern).each {
-            def name = it.first()
-            // Handle the case of "Jan och Maria Larsson"
-            def previousName = names.isEmpty() ? null : names.last()
-            if (previousName?.split()?.size() == 1 && s =~ /$previousName$conjPattern$name/) {
-                def nameParts = name.split()
-                if (nameParts.size() > 1) {
-                    names[-1] += " ${nameParts.last()}"
-                }
-            }
-            names << name
-        }
-
-        return names
     }
 }
