+++ whelktool/src/main/groovy/datatool/scripts/mergeworks/WorkToolJob.groovy
@@ -18,13 +18,8 @@ import java.util.function.Function
 
 import static datatool.scripts.mergeworks.FieldStatus.DIFF
 
-import static datatool.scripts.mergeworks.Util.asList
-import static datatool.scripts.mergeworks.Util.chipString
 import static datatool.scripts.mergeworks.Util.getPathSafe
-import static datatool.scripts.mergeworks.Util.normalize
 import static datatool.scripts.mergeworks.Util.partition
-import static datatool.scripts.mergeworks.Util.parseRespStatement
-import static datatool.scripts.mergeworks.Util.Relator
 
 class WorkToolJob {
     Whelk whelk
@@ -403,332 +398,6 @@ class WorkToolJob {
                 }
             }
         })
-    }
-
-    void add9pu() {
-        statistics.printOnShutdown()
-        run({ cluster ->
-            return {
-                statistics.increment('add 9pu', 'clusters checked')
-                def docs = cluster
-                        .collect(whelk.&getDocument)
-                        .findAll()
-                        .collect { [doc: it, checksum: it.getChecksum(whelk.jsonld), changed: false] }
-
-                def ill = ['@id': Relator.ILLUSTRATOR.iri]
-                def pu = ['@id': Relator.PRIMARY_RIGHTS_HOLDER.iri]
-                def path = ['@graph', 1, 'instanceOf', 'contribution']
-
-                docs.each {
-                    Document d = it.doc
-
-                    statistics.increment('add 9pu', 'docs checked')
-
-                    getPathSafe(d.data, path, []).each { Map c ->
-                        def r = asList(c.role)
-
-                        if (pu in r || !(ill in r) || c.'@type' == 'PrimaryContribution')
-                            return
-
-                        for (Map other : docs) {
-                            Document od = other.doc
-
-                            def found9pu = false
-
-                            getPathSafe(od.data, path, []).each { Map oc ->
-                                if (asList(c.agent) == asList(oc.agent) && asList(oc.role).containsAll([ill, pu])) {
-                                    c.role = asList(c.role) + pu
-                                    found9pu = true
-                                    statistics.increment('add 9pu', "9pu added")
-                                    if (verbose) {
-                                        println("${d.shortId} <- ${od.shortId}")
-                                    }
-                                    return
-                                }
-                            }
-
-                            if (found9pu) {
-                                println(c)
-                                it.changed = true
-                                break
-                            }
-                        }
-                    }
-                }
-
-                docs.each {
-                    if (!dryRun && it.changed) {
-                        Document d = it.doc
-                        d.setGenerationDate(new Date())
-                        d.setGenerationProcess(generationProcess)
-                        whelk.storeAtomicUpdate(d, !loud, changedIn, changedBy, it.checksum)
-                    }
-                }
-            }
-        })
-    }
-
-    void fetchContributionFromRespStatement() {
-        def loadThingByIri = { String iri ->
-            // TODO: fix whelk, add load by IRI method
-            whelk.storage.loadDocumentByMainId(iri)?.with { doc ->
-                return (Map) doc.data['@graph'][1]
-            }
-        }
-
-        def loadIfLink = { it['@id'] ? loadThingByIri(it['@id']) : it }
-
-        statistics.printOnShutdown()
-        run({ cluster ->
-            return {
-                statistics.increment('fetch contribution from respStatement', 'clusters checked')
-                def docs = cluster
-                        .collect(whelk.&getDocument)
-                        .findAll()
-                        .collect { [doc: it, checksum: it.getChecksum(whelk.jsonld), changed: false] }
-
-                docs.each {
-                    Document d = it.doc
-                    def respStatement = getPathSafe(d.data, ['@graph', 1, 'responsibilityStatement'])
-                    if (!respStatement)
-                        return
-
-                    statistics.increment('fetch contribution from respStatement', 'docs checked')
-
-                    def contributionsInRespStmt = parseRespStatement(respStatement)
-                    def contribution = getPathSafe(d.data, ['@graph', 1, 'instanceOf', 'contribution'], [])
-
-                    contribution.each { Map c ->
-                        asList(c.agent).each { a ->
-                            def matchedOnName = contributionsInRespStmt.find { n, r ->
-                                nameMatch(n, loadIfLink(a))
-                            }
-
-                            if (!matchedOnName)
-                                return
-
-                            // Contributor found locally, omit from further search
-                            contributionsInRespStmt.remove(matchedOnName.key)
-
-                            def dontAdd = { Relator relator, boolean isFirstStmtPart ->
-                                relator == Relator.UNSPECIFIED_CONTRIBUTOR
-                                        || isFirstStmtPart && relator == Relator.AUTHOR
-                                        && c.'@type' != 'PrimaryContribution'
-                            }
-
-                            def rolesInRespStatement = matchedOnName.value
-                                    .findResults { dontAdd(it) ? null : it.getV1() }
-
-                            if (rolesInRespStatement.isEmpty())
-                                return
-
-                            def rolesInContribution = asList(c.role).findAll { it.'@id' != Relator.UNSPECIFIED_CONTRIBUTOR.iri }
-
-                            // Replace Adapter with Editor
-                            it.changed |= rolesInRespStatement.removeAll { r ->
-                                r == Relator.EDITOR && rolesInContribution.findIndexOf {
-                                    it.'@id' == Relator.ADAPTER.iri
-                                }.with {
-                                    if (it == -1) {
-                                        return false
-                                    } else {
-                                        rolesInContribution[it]['@id'] = Relator.EDITOR.iri
-                                        return true
-                                    }
-                                }
-                            }
-
-                            if (rolesInRespStatement.size() <= rolesInContribution.size())
-                                return
-
-                            rolesInRespStatement.each { r ->
-                                def idLink = ['@id': r.iri]
-                                if (!(idLink in rolesInContribution)) {
-                                    rolesInContribution << idLink
-                                    it.changed = true
-                                    def roleShort = r.iri.split('/').last()
-                                    statistics.increment('fetch contribution from respStatement', "$roleShort roles specified")
-                                    if (verbose) {
-                                        println("${chipString(c, whelk)} (${d.shortId}) <- $roleShort")
-                                    }
-                                }
-                            }
-
-                            c.role = rolesInContribution
-                        }
-                    }
-
-                    def comparable = {
-                        it*.getV1().findResults { Relator r ->
-                            r != Relator.UNSPECIFIED_CONTRIBUTOR
-                                    ? ['@id': r.iri]
-                                    : null
-                        }
-                    }
-
-                    contributionsInRespStmt.each { name, roles ->
-                        for (Map other : docs) {
-                            Document od = other.doc
-                            def matched = getPathSafe(od.data, ['@graph', 1, 'instanceOf', 'contribution'], [])
-                                    .find { Map c ->
-                                        asList(c.agent).any { a ->
-                                            loadIfLink(a).with { nameMatch(name, it) && !(it.description =~ /(?i)pseud/) }
-                                                    && comparable(roles).with { r -> !r.isEmpty() && asList(c.role).containsAll(r) }
-                                                    && Util.bestEncodingLevel.indexOf(d.getEncodingLevel()) <= Util.bestEncodingLevel.indexOf(od.getEncodingLevel())
-                                        }
-                                    }
-                            if (matched) {
-                                contribution << matched
-                                roles.each {
-                                    def roleShort = it.getV1().iri.split('/').last()
-                                    statistics.increment('fetch contribution from respStatement', "$roleShort found in cluster")
-                                }
-                                if (verbose) {
-                                    println("${d.shortId} <- ${chipString(matched, whelk)} (${od.shortId})")
-                                }
-                                it.changed = true
-                                break
-                            }
-                        }
-                    }
-                }
-
-                docs.each {
-                    if (!dryRun && it.changed) {
-                        Document d = it.doc
-                        d.setGenerationDate(new Date())
-                        d.setGenerationProcess(generationProcess)
-                        whelk.storeAtomicUpdate(d, !loud, changedIn, changedBy, it.checksum)
-                    }
-                }
-            }
-        }
-
-        )
-    }
-
-    void linkContribution() {
-        def loadThingByIri = { String iri ->
-            // TODO: fix whelk, add load by IRI method
-            whelk.storage.loadDocumentByMainId(iri)?.with { doc ->
-                return (Map) doc.data['@graph'][1]
-            }
-        }
-
-        def loadIfLink = { it['@id'] ? loadThingByIri(it['@id']) : it }
-
-        statistics.printOnShutdown()
-        run({ cluster ->
-            return {
-                statistics.increment('link contribution', 'clusters checked')
-                // TODO: check work language?
-                def docs = cluster
-                        .collect(whelk.&getDocument)
-                        .collect { [doc: it, checksum: it.getChecksum(whelk.jsonld), changed: false] }
-
-                List<Map> linked = []
-                docs.each { d ->
-                    def contribution = getPathSafe(d.doc.data, ['@graph', 1, 'instanceOf', 'contribution'], [])
-                    contribution.each { Map c ->
-                        if (c.agent && c.agent['@id']) {
-                            loadThingByIri(c.agent['@id'])?.with { Map agent ->
-                                agent.roles = asList(c.role)
-                                linked << agent
-                            }
-                        }
-                    }
-                    statistics.increment('link contribution', 'docs checked')
-                }
-
-                docs.each {
-                    Document d = it.doc
-                    def contribution = getPathSafe(d.data, ['@graph', 1, 'instanceOf', 'contribution'], [])
-                    contribution.each { Map c ->
-                        if (c.agent && !c.agent['@id']) {
-                            def l = linked.find {
-                                agentMatches(c.agent, it) && (!c.role || it.roles.containsAll(c.role))
-                            }
-                            if (l) {
-                                println("${d.shortId} ${chipString(c, whelk)} --> ${chipString(l, whelk)}")
-                                c.agent = ['@id': l['@id']]
-                                it.changed = true
-                                statistics.increment('link contribution', 'agents linked')
-                            } else if (verbose) {
-                                println("${d.shortId} NO MATCH: ${chipString(c, whelk)} ??? ${linked.collect { chipString(it, whelk) }}")
-                            }
-                        }
-                    }
-                }
-
-                List<Map> primaryAutAgents = []
-                docs.each {
-                    def contribution = getPathSafe(it.doc.data, ['@graph', 1, 'instanceOf', 'contribution'], [])
-                    def p = contribution.findAll()
-                    contribution.each {
-                        if (it['@type'] == 'PrimaryContribution' && it['role'] == ['@id': 'https://id.kb.se/relator/author'] && it['agent']) {
-                            Map agent = loadIfLink(it['agent'])
-                            if (agent) {
-                                primaryAutAgents << agent
-                            }
-                        }
-                    }
-                }
-
-                docs.each {
-                    Document d = it.doc
-                    def contribution = getPathSafe(d.data, ['@graph', 1, 'instanceOf', 'contribution'], [])
-                    contribution.each { Map c ->
-                        if (c['@type'] == 'PrimaryContribution' && !c.role) {
-                            if (c.agent) {
-                                def agent = loadIfLink(c.agent)
-                                if (primaryAutAgents.any { agentMatches(agent, it) }) {
-                                    c.role = ['@id': 'https://id.kb.se/relator/author']
-                                    it.changed = true
-                                    statistics.increment('link contribution', 'author role added to primary contribution')
-                                }
-                            }
-                        }
-                    }
-                }
-
-                docs.each {
-                    if (!dryRun && it.changed) {
-                        Document d = it.doc
-                        d.setGenerationDate(new Date())
-                        d.setGenerationProcess(generationProcess)
-                        whelk.storeAtomicUpdate(d, !loud, changedIn, changedBy, it.checksum)
-                    }
-                }
-            }
-        })
-    }
-
-    static boolean agentMatches(Map local, Map linked) {
-        nameMatch(local, linked) && !yearMismatch(local, linked)
-    }
-
-    static boolean nameMatch(Object local, Map agent) {
-        def variants = [agent] + asList(agent.hasVariant)
-        def name = {
-            Map p ->
-                (p.givenName && p.familyName)
-                        ? normalize("${p.givenName} ${p.familyName}")
-                        : p.name ? normalize("${p.name}") : null
-        }
-
-        def localName = local instanceof Map ? name(local) : normalize(local)
-
-        localName && variants.any {
-            name(it) && localName == name(it)
-        }
-    }
-
-    static boolean yearMismatch(Map local, Map linked) {
-        def birth = { Map p -> p.lifeSpan?.with { (it.replaceAll(/[^\-0-9]/, '').split('-') as List)[0] } }
-        def death = { Map p -> p.lifeSpan?.with { (it.replaceAll(/[^\-0-9]/, '').split('-') as List)[1] } }
-        def b = birth(local) && birth(linked) && birth(local) != birth(linked)
-        def d = death(local) && death(linked) && death(local) != death(linked)
-        b || d
     }
 
     private void run(Function<List<String>, Runnable> f) {
